crnn/requirements.txt
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
eval_type_backport==0.2.0
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-python-headless==4.10.0.84
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.5.0
simsimd==6.0.5
stringzilla==3.10.10
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
